Use char buffers instead of stdio hacks (breaks TOC stuff)
This is quite a bit cleaner than using the non-POSIX stdio
functions (funopen/fopencookie) and lets us run on Solaris
and Windows (maybe).

The TOC features are not yet available without the stdio hacks
and have been disabled temporarily. A new version of Discount
should be available soon that remedies this.