Recognize htmlblock close tags anywhere in line
Previously htmlblock() could only detect a closing tag if it was on
its own on a line. Unfortunately the Pygments library when generating
code blocks with line numbers does not put the closing </table> on
its own line. Instead it writes it as </td></tr></table>, in this
form the Discount library would not detect the closing tag and would
interpret the remainder of the document as HTML.

Signed-off-by: Ryan Tomayko <rtomayko@gmail.com>

NOTE: The original patch included modifications for the Markdown
conformance suite. I didn't want to modify the conformance suite
directly so I created a simple test case in test/markdown_test.rb